Ass#4  |  Refactor "isSelected" functionality.

+++ src/a4/objects/Car.java
@@ -525,8 +525,8 @@ public class Car extends Moveable implements ISteerable , IDrawable, ICollider{
     @Override
     public boolean collidesWith(ICollider obj) {
 
-        float distX = this.getX() - ((GameObject)obj).getX();
-        float distY = this.getY() - ((GameObject)obj).getY();
+        double distX = this.getX() - ((GameObject)obj).getX();
+        double distY = this.getY() - ((GameObject)obj).getY();
         float distanceBtwnCenters = (float) Math.sqrt(distX * distX + distY * distY);
 
         if((this.getDistanceOfReference() + obj.getDistanceOfReference() >
