+++ src/a4/objects/NPCCar.java
@@ -194,8 +194,8 @@ public class NPCCar extends Car {
      */
     @Override
     public boolean collidesWith(ICollider obj) {
-        float distX = this.getX() - ((GameObject)obj).getX();
-        float distY = this.getY() - ((GameObject)obj).getY();
+        double distX = this.getX() - ((GameObject)obj).getX();
+        double distY = this.getY() - ((GameObject)obj).getY();
         float distanceBtwnCenters = (float) Math.sqrt(distX * distX + distY * distY);
 
         if((this.getDistanceOfReference() + obj.getDistanceOfReference() >
